Use better Jira links
Currently the "create an issue" link takes you to the RHEL project in
Jira.  Users then have to locate the "Create" button at the top of the
page, make a decision on issue type (epic/story/bug), and then fill out
the "Affects Version/s" field appropriately.  This is a lot of friction.

Instead, this change adds separate links for EL8, EL9, and EL10, with
the issue type set to bug, and the "Affects Version/s" field
pre-populated.